port remaining plugin nodes
